bot/requirements.txt
pipecat-ai[deepgram,cartesia,silero,twilio,anthropic]>=0.0.108
anthropic>=0.40.0
deepgram-sdk>=3.0.0
cartesia>=1.0.0
twilio>=9.0.0
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
websockets>=12.0
aiohttp>=3.9.0
python-dotenv>=1.0.0
loguru>=0.7.0
